# schema/openapi/client-api.yaml
openapi: 3.1.0
info:
  title: Tima Client API
  version: 1.0.0-phase1-slice1
  description: |
    User-facing REST contract. Bot runtime, WebSocket, operations, internal and legal APIs are excluded.
    All timestamps are RFC 3339 UTC. All IDs are UUIDs unless a schema says otherwise.
    Successful and error responses may include X-RateLimit-Limit, X-RateLimit-Remaining and
    X-RateLimit-Reset. A 429 response also includes Retry-After.
  x-tima-implementation-phase: phase-0
...
servers:
  - url: https://api.tima.example/v1
security:
  - bearerAuth: []
    deviceId: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: startRegistration
      security: []
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/RegisterStart'}
      responses:
        '202': {description: OTP sent, content: {application/json: {schema: {$ref: '#/components/schemas/OtpChallenge'}}}}
        default: {$ref: '#/components/responses/Error'}
  /auth/verify:
    post:
      tags: [Auth]
      operationId: verifyRegistration
      security: []
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/AttestationToken'}]
      requestBody: {$ref: '#/components/requestBodies/RegisterVerify'}
      responses:
        '201': {description: Account and session created, content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}}
        default: {$ref: '#/components/responses/Error'}
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/AttestationToken'}]
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses:
        '200': {description: Authenticated, content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}}
        default: {$ref: '#/components/responses/Error'}
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security: []
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/Refresh'}
      responses:
        '200': {description: Tokens rotated, content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}}
        default: {$ref: '#/components/responses/Error'}
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      x-tima-implementation-phase: phase-1
      responses:
        '204': {description: Session revoked}
        default: {$ref: '#/components/responses/Error'}
  /verify/attestation/ios:
    post:
      tags: [Auth]
      operationId: verifyIosAttestation
      security: []
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/IosAttestation'}
      responses:
        '200': {description: Attestation accepted, content: {application/json: {schema: {$ref: '#/components/schemas/AttestationResult'}}}}
        default: {$ref: '#/components/responses/Error'}
  /verify/integrity/android:
    post:
      tags: [Auth]
      operationId: verifyAndroidIntegrity
      security: []
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/AndroidIntegrity'}
      responses:
        '200': {description: Integrity accepted, content: {application/json: {schema: {$ref: '#/components/schemas/AttestationResult'}}}}
        default: {$ref: '#/components/responses/Error'}
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      x-tima-implementation-phase: phase-1
      responses:
        '200': {description: Current profile, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}
        default: {$ref: '#/components/responses/Error'}
    patch:
      tags: [Users]
      operationId: updateCurrentUser
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/UserPatch'}
      responses:
        '200': {description: Updated profile, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}
        default: {$ref: '#/components/responses/Error'}
  /devices:
    get:
      tags: [Users]
      operationId: listDevices
      x-tima-implementation-phase: phase-1
      responses:
        '200': {description: Active sessions, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceList'}}}}
        default: {$ref: '#/components/responses/Error'}
  /devices/{id}:
    delete:
      tags: [Users]
      operationId: revokeDevice
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '204': {description: Device revoked}
        default: {$ref: '#/components/responses/Error'}
  /devices/push:
    put:
      tags: [Users]
      operationId: registerDevicePush
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/PushRegistration'}
      responses:
        '204': {description: Push registration stored}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Users]
      operationId: deleteDevicePush
      x-tima-implementation-phase: phase-1
      responses:
        '204': {description: Push registration removed}
        default: {$ref: '#/components/responses/Error'}
  /link/session:
    post:
      tags: [Users]
      operationId: startDeviceLink
      security: []
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/LinkSession'}
      responses:
        '201': {description: Link session, content: {application/json: {schema: {$ref: '#/components/schemas/LinkChallenge'}}}}
        default: {$ref: '#/components/responses/Error'}
  /link/confirm:
    post:
      tags: [Users]
      operationId: confirmDeviceLink
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/LinkConfirm'}
      responses:
        '200': {description: Linked device, content: {application/json: {schema: {$ref: '#/components/schemas/Device'}}}}
        default: {$ref: '#/components/responses/Error'}
  /link/claim:
    post:
      tags: [Users]
      operationId: claimDeviceLink
      security: []
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/LinkClaim'}
      responses:
        '200': {description: Desktop session and wrapped device secret, content: {application/json: {schema: {$ref: '#/components/schemas/LinkedAuthSession'}}}}
        default: {$ref: '#/components/responses/Error'}
  /keys/bundle:
    put:
      tags: [Keys]
      operationId: putKeyBundle
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/KeyBundle'}
      responses:
        '200': {description: Bundle stored, content: {application/json: {schema: {$ref: '#/components/schemas/KeyBundle'}}}}
        default: {$ref: '#/components/responses/Error'}
  /keys/bundle/{user_id}:
    get:
      tags: [Keys]
      operationId: getKeyBundle
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses:
        '200': {description: Device bundles and consumed one-time prekey, content: {application/json: {schema: {$ref: '#/components/schemas/UserKeyBundles'}}}}
        default: {$ref: '#/components/responses/Error'}
  /keys/prekeys/replenish:
    post:
      tags: [Keys]
      operationId: replenishPrekeys
      x-tima-implementation-phase: phase-1
      requestBody: {$ref: '#/components/requestBodies/Prekeys'}
      responses:
        '200': {description: Prekeys stored, content: {application/json: {schema: {$ref: '#/components/schemas/PrekeyResult'}}}}
        default: {$ref: '#/components/responses/Error'}
  /escrow/config:
    get:
      tags: [Keys]
      operationId: getEscrowConfig
      x-tima-implementation-phase: phase-1
      parameters:
        - {$ref: '#/components/parameters/ConversationType'}
        - {$ref: '#/components/parameters/ConversationId'}
        - {$ref: '#/components/parameters/Epoch'}
        - {$ref: '#/components/parameters/Shard'}
      responses:
        '200': {description: Signed current and next regional public configuration, content: {application/json: {schema: {$ref: '#/components/schemas/EscrowConfig'}}}}
        default: {$ref: '#/components/responses/Error'}
  /chats:
    get:
      tags: [Chats]
      operationId: listChats
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Chats, content: {application/json: {schema: {$ref: '#/components/schemas/ChatList'}}}}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Chats]
      operationId: createChat
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/ChatCreate'}
      responses:
        '201': {description: Chat created, content: {application/json: {schema: {$ref: '#/components/schemas/Chat'}}}}
        default: {$ref: '#/components/responses/Error'}
  /chats/{id}/messages:
    get:
      tags: [Chats]
      operationId: listChatMessages
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Private message history. Each message includes only the wrapped key authorized for the authenticated X-Device-Id; wrapped keys for other devices are never returned., content: {application/json: {schema: {$ref: '#/components/schemas/PrivateMessageList'}}}}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Chats]
      operationId: sendChatMessage
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/PrivateMessage'}
      responses:
        '201': {description: Private message accepted, content: {application/json: {schema: {$ref: '#/components/schemas/PrivateMessage'}}}}
        default: {$ref: '#/components/responses/Error'}
  /chats/{id}/message-reservations:
    post:
      tags: [Chats]
      operationId: reserveChatMessage
      security:
        - bearerAuth: []
          deviceId: []
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '201': {description: Server-assigned identifiers reserved for canonical message signing., content: {application/json: {schema: {$ref: '#/components/schemas/MessageReservation'}}}}
        default: {$ref: '#/components/responses/Error'}
  /chats/{id}/messages/{msg_id}:
    delete:
      tags: [Chats]
      operationId: deleteChatMessage
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/MessageId'}]
      responses:
        '204': {description: Message soft-deleted}
        default: {$ref: '#/components/responses/Error'}
  /chats/{id}/messages/{msg_id}/revisions:
    post:
      tags: [Chats]
      operationId: reviseChatMessage
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/MessageId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/PrivateMessage'}
      responses:
        '201': {description: Immutable private revision, content: {application/json: {schema: {$ref: '#/components/schemas/PrivateRevision'}}}}
        default: {$ref: '#/components/responses/Error'}
  /chats/{id}/read:
    post:
      tags: [Chats]
      operationId: markChatRead
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/ReadCursor'}
      responses:
        '200': {description: Read cursor stored, content: {application/json: {schema: {$ref: '#/components/schemas/ReadState'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /chats/{chat_id}/media/uploads:
    post:
      tags: [Media]
      operationId: createChatMediaUpload
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/ChatId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/MediaUploadCreate'}
      responses:
        '201': {description: Private upload slots, content: {application/json: {schema: {$ref: '#/components/schemas/MediaUpload'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /posts/assets:
    post:
      tags: [Media, Posts]
      operationId: createPostAssetUpload
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/MediaUploadCreate'}
      responses:
        '201': {description: Public staging upload slots, content: {application/json: {schema: {$ref: '#/components/schemas/MediaUpload'}}}}
        default: {$ref: '#/components/responses/Error'}
  /media/uploads/{media_id}/complete:
    post:
      tags: [Media]
      operationId: completeMediaUpload
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/MediaId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/MediaComplete'}
      responses:
        '202': {description: Scan and processing started, content: {application/json: {schema: {$ref: '#/components/schemas/Media'}}}}
        default: {$ref: '#/components/responses/Error'}
  /media/{media_id}/access:
    post:
      tags: [Media]
      operationId: createMediaAccess
      x-tima-implementation-phase: phase-1
      parameters: [{$ref: '#/components/parameters/MediaId'}]
      requestBody: {$ref: '#/components/requestBodies/MediaAccess'}
      responses:
        '200': {description: Variant URL valid for 15 minutes, content: {application/json: {schema: {$ref: '#/components/schemas/MediaAccess'}}}}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Device-bound user access token.
    deviceId:
      type: apiKey
      in: header
      name: X-Device-Id
      description: UUID of the authenticated device.
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    AttestationToken:
      name: X-Attestation-Token
      in: header
      required: false
      description: Required for mobile clients according to platform policy.
      schema: {type: string, minLength: 16, maxLength: 8192}
...
  responses:
    Error:
      description: Validation, authentication, authorization, absence, conflict, rate-limit or dependency failure.
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        X-RateLimit-Limit: {$ref: '#/components/headers/RateLimitLimit'}
        X-RateLimit-Remaining: {$ref: '#/components/headers/RateLimitRemaining'}
        X-RateLimit-Reset: {$ref: '#/components/headers/RateLimitReset'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  requestBodies:
    RegisterStart: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterStart'}}}}
    RegisterVerify: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterVerify'}}}}
    Login: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Login'}}}}
    Refresh: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Refresh'}}}}
    IosAttestation: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IosAttestation'}}}}
    AndroidIntegrity: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AndroidIntegrity'}}}}
    UserPatch: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserPatch'}}}}
    LinkSession: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LinkSession'}}}}
    LinkConfirm: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LinkConfirm'}}}}
...
    PushRegistration: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PushRegistration'}}}}
...
  schemas:
    Uuid: {type: string, format: uuid}
...
    OtpChallenge:
      type: object
      additionalProperties: false
      required: [challenge_id, expires_at, resend_after]
      properties:
        challenge_id: {$ref: '#/components/schemas/Uuid'}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
        resend_after: {type: integer, minimum: 0}
...
    AuthSession:
      type: object
      additionalProperties: false
      required: [access_token, refresh_token, token_type, expires_in, user, device]
      properties:
        access_token: {type: string, minLength: 32}
        refresh_token: {type: string, minLength: 32}
        token_type: {type: string, const: Bearer}
        expires_in: {type: integer, minimum: 1}
        user: {$ref: '#/components/schemas/User'}
        device: {$ref: '#/components/schemas/Device'}
...
    AttestationResult:
      type: object
      additionalProperties: false
      required: [attestation_token, expires_at, verdict]
      properties:
        attestation_token: {type: string, minLength: 16}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
        verdict: {type: string, enum: [trusted, limited]}
    User:
      type: object
      additionalProperties: false
      required: [id, account_type, display_name, account_home_region, created_at]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        account_type: {type: string, enum: [human, virtual]}
        username: {type: string, pattern: '^[a-zA-Z0-9_]{3,32}$'}
        display_name: {type: string, minLength: 1, maxLength: 100}
        bio: {type: string, maxLength: 1000}
        avatar_media_id: {$ref: '#/components/schemas/Uuid'}
        account_home_region: {$ref: '#/components/schemas/Region'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
...
    DeviceList:
      type: object
      required: [items]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/Device'}}}
...
    LinkChallenge:
      type: object
      additionalProperties: false
      required: [session_id, qr_payload, claim_token, expires_at]
      properties:
        session_id: {$ref: '#/components/schemas/Uuid'}
        qr_payload: {type: string, minLength: 16}
        claim_token: {type: string, minLength: 32}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
...